examples/taxis/requirements.txt
dill
kafka-python
psycopg[binary]
pandas
requests
river
